# Documents/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Farm API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Cows": {
      "get": {
        "tags": [
          "Cows"
        ],
        "summary": "Get list of cows",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CowDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CowDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CowDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Cows/{id}": {
      "get": {
        "tags": [
          "Cows"
        ],
        "summary": "Get Cow based on cow ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cow Id",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Cow Id",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CowDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CowDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CowDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Cows/{cowId}": {
      "put": {
        "tags": [
          "Cows"
        ],
        "summary": "Update the cow's status based on CowId",
        "parameters": [
          {
            "name": "cowId",
            "in": "path",
            "description": "cowId to update cow",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "cowId to update cow",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Dto to update status",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CowDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CowDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CowDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Cows/count": {
      "get": {
        "tags": [
          "Cows"
        ],
        "summary": "Get Cow count base on farm , state and Date",
        "parameters": [
          {
            "name": "farm",
            "in": "query",
            "description": "farm",
            "schema": {
              "type": "string",
              "description": "farm",
              "nullable": true
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "state like Open, Inseminated, Pregnant, Dry",
            "schema": {
              "type": "string",
              "description": "state like Open, Inseminated, Pregnant, Dry",
              "nullable": true
            }
          },
          {
            "name": "onDate",
            "in": "query",
            "description": "onDate",
            "schema": {
              "type": "string",
              "description": "onDate",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "CowDto": {
        "type": "object",
        "properties": {
          "cowId": {
            "type": "integer",
            "format": "int32"
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "farm": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StateDto": {
        "required": [
          "state"
        ],
        "type": "object",
        "properties": {
          "state": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}